+++ src/__main__.py
@@ -1,9 +1,7 @@
 import sys
 
 from PySide6.QtWidgets import QApplication
-from src.TTSengine.TTSmain import MainWindow
-
-# TODO: Redo the file structure, move all the TTS engine stuff to a separate folder
+from src.gui.mainWindow import MainWindow
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
