+++ src/gui/ui/configwindow.ui
